Updated octree

+++ oct_illustris/octree.py
@@ -21,16 +21,16 @@ class octree(object):
 		return self._height
 
 	def build(self):
-		# By setting dtype to int64, the maximum level is 20 (2^(3*20)<2^63-1)
+		# By setting dtype to uint32, the maximum level is 10 (2^(3*10)<2^32-1)
 		idx_3d = np.astype(
 			2**self._level * (self._pos - self._boundary[0]) / 
-			(self._boundary[1] - self._boundary[0]), dtype='i8')
+			(self._boundary[1] - self._boundary[0]), dtype="u4")
 
 		# Conbine 3D index into 1D 
 		idx_1d = np.left_shift(idx_3d, [2*self._height,self._height,0])
 
 		# Sort rank with index
-		rank = np.arange(self._offset, self._offset+self._length, dtype='i8')
+		rank = np.arange(self._offset, self._offset+self._length, dtype="u4")
 		return rank[np.argsort(idx_1d)]
